+++ CHANGELOG.md
@@ -16,11 +16,11 @@ We do not give estimated times for completion on `Accepted` Proposals.
 
 ---
 
-## [Unreleased] - 2019-09-17
+## [v5.9.0] - 2019-09-17
 
 `Added`
 
-- Custom interfaces, traits and uses.
+- Custom interfaces, traits, uses and phpDoc tags type and description.
 
 ## [v5.8.2] - 2019-09-17
 
@@ -128,7 +128,8 @@ It can work with `nightly`, but is not compatible with `hhvm`, as it uses differ
 [Accepted]: https://github.com/Triun/laravel-model-base/labels/Accepted
 [Rejected]: https://github.com/Triun/laravel-model-base/labels/Rejected
 
-[Unreleased]: https://github.com/Triun/laravel-model-base/compare/v5.8.2...master
+[Unreleased]: https://github.com/Triun/laravel-model-base/compare/v5.9.0...master
+[v5.9.0]: https://github.com/Triun/laravel-model-base/compare/v5.8.2...v5.9.0
 [v5.8.2]: https://github.com/Triun/laravel-model-base/compare/v5.8.1...v5.8.2
 [v5.8.1]: https://github.com/Triun/laravel-model-base/compare/v5.8.0...v5.8.1
 [v5.8.0]: https://github.com/Triun/laravel-model-base/compare/v5.7.0...v5.8.0